update gradle,edit settings page
