+++ .gitignore
@@ -1,9 +1,10 @@
 *.iml
 .gradle
 /local.properties
+/.idea
 /.idea/workspace.xml
 /.idea/libraries
-.DS_Store
+.DS_Store/
 /build
 /captures
 .externalNativeBuild
