+++ app/build.gradle
@@ -14,7 +14,7 @@ def buildTime() {
 
 android {
     compileSdkVersion 25
-    buildToolsVersion "25.0.3"
+    buildToolsVersion '27.0.3'
     defaultConfig {
         applicationId "org.zky.tool.magnetsearch"
         minSdkVersion 19
@@ -47,12 +47,12 @@ android {
 
             applicationVariants.all {
                 variant ->
-                    variant.outputs.each {
+                    variant.outputs.all {
                         output ->
                             def outputFile = output.outputFile
                             if (outputFile != null && outputFile.name.endsWith('.apk')) {
                                 def fileName = "MagnetSearch_v${defaultConfig.versionName}.apk"
-                                output.outputFile = new File('/Users/zhangkun/Documents/apk', fileName)
+                                outputFileName = new File('apk', fileName)
                                 //for windows
 //                                output.outputFile = new File('C:/apk', fileName)
 
@@ -64,11 +64,17 @@ android {
 
     }
 
+    flavorDimensions "market"
+
     productFlavors {
         play {
+            dimension "market"
+
             versionName "1.3.2.${buildTime()}${gitVersion}-play"
         }
         kuan{
+            dimension "market"
+
             versionName "1.3.2.${buildTime()}${gitVersion}-kuan"
         }
 
@@ -98,33 +104,33 @@ greendao {
 }
 
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
+    api fileTree(include: ['*.jar'], dir: 'libs')
+    androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
-    compile "com.android.support:appcompat-v7:$supportLibrary"
-    compile "com.android.support:cardview-v7:$supportLibrary"
-    compile "com.android.support:design:$supportLibrary"
-    compile 'com.lzp.floatingactionbutton:floatingactionbuttonplus:1.0.0'
-    compile 'com.android.support.constraint:constraint-layout:1.0.0-alpha6'
-    compile 'com.jakewharton:butterknife:8.4.0'
-    compile 'com.google.code.gson:gson:2.8.0'
-    compile 'io.reactivex:rxjava:1.1.0'
-    compile 'io.reactivex:rxandroid:1.1.0'
-    compile 'com.squareup.retrofit2:retrofit:2.1.0'
-    compile 'com.squareup.retrofit2:adapter-rxjava:2.1.0'
-    compile 'com.squareup.retrofit2:converter-gson:2.0.0-beta4'
-    compile 'org.greenrobot:greendao:3.0.1'
-    compile 'org.greenrobot:greendao-generator:3.0.0'
-    testCompile 'junit:junit:4.12'
+    implementation "com.android.support:appcompat-v7:$supportLibrary"
+    implementation "com.android.support:cardview-v7:$supportLibrary"
+    implementation "com.android.support:design:$supportLibrary"
+    implementation 'com.lzp.floatingactionbutton:floatingactionbuttonplus:1.0.0'
+    implementation 'com.android.support.constraint:constraint-layout:1.0.0-alpha6'
+    implementation 'com.jakewharton:butterknife:8.4.0'
+    implementation 'com.google.code.gson:gson:2.8.0'
+    implementation 'io.reactivex:rxjava:1.1.0'
+    implementation 'io.reactivex:rxandroid:1.1.0'
+    implementation 'com.squareup.retrofit2:retrofit:2.1.0'
+    implementation 'com.squareup.retrofit2:adapter-rxjava:2.1.0'
+    implementation 'com.squareup.retrofit2:converter-gson:2.0.0-beta4'
+    implementation 'org.greenrobot:greendao:3.0.1'
+    implementation 'org.greenrobot:greendao-generator:3.0.0'
+    testImplementation 'junit:junit:4.12'
     annotationProcessor 'com.jakewharton:butterknife-compiler:8.4.0'
-    compile 'com.google.zxing:core:3.3.0'
-    compile 'org.jsoup:jsoup:1.10.2'
-    compile 'com.google.firebase:firebase-core:9.6.1'
-    compile 'com.google.firebase:firebase-ads:9.6.1'
-    compile 'com.google.firebase:firebase-crash:9.6.1'
-    compile 'com.google.android.gms:play-services:9.6.1'
-    compile 'org.mod4j.org.apache.commons:lang:2.1.0'
+    implementation 'com.google.zxing:core:3.3.0'
+    implementation 'org.jsoup:jsoup:1.10.2'
+    implementation 'com.google.firebase:firebase-core:9.6.1'
+    implementation 'com.google.firebase:firebase-ads:9.6.1'
+    implementation 'com.google.firebase:firebase-crash:9.6.1'
+    implementation 'com.google.android.gms:play-services:9.6.1'
+    implementation 'org.mod4j.org.apache.commons:lang:2.1.0'
 }
 
 apply plugin: 'com.google.gms.google-services'
