+++ build.gradle
@@ -7,9 +7,10 @@ buildscript {
             url 'https://maven.google.com'
         }
 
+        google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.3.3'
+        classpath 'com.android.tools.build:gradle:3.1.4'
         classpath 'org.greenrobot:greendao-gradle-plugin:3.0.0'
         classpath 'com.google.gms:google-services:3.0.0'
 
@@ -22,6 +23,7 @@ allprojects {
     repositories {
         jcenter()
         mavenCentral()
+        google()
     }
 }
